Added Measurements class
Added a module to make measurements of the following for the germ band
and the growth zone:
cell count
length
area
and average cell size for the germ band.

The module currently outputs to the command line, but the functions
could be used to write the measurements to a file for analysis.